+++ packages/cryptoplease/lib/features/app_lock/routes.dart
@@ -1,8 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 
-import 'src/presentation/app_lock_disable_screen.dart';
-import 'src/presentation/app_lock_enable_screen.dart';
-import 'src/presentation/app_lock_setup_flow_screen.dart';
+import 'src/widgets/app_lock_disable_screen.dart';
+import 'src/widgets/app_lock_enable_screen.dart';
+import 'src/widgets/app_lock_setup_flow_screen.dart';
 
 const appLockSetupFlowRoutes = AutoRoute<void>(
   page: AppLockSetupFlowScreen,
