+++ packages/cryptoplease/lib/features/app_lock/src/widgets/components/pin_display.dart
