+++ packages/cryptoplease/lib/features/app_lock/src/widgets/components/switch.dart
